+++ provdbconnector/db_adapters/neo4j/cypher_commands.py
@@ -1,0 +1,82 @@
+
+NEO4J_TEST_CONNECTION = """MATCH (n) RETURN count(n) as count"""
+
+# create
+NEO4J_CREATE_DOCUMENT_NODE_RETURN_ID = """CREATE (node { }) RETURN ID(node) as ID"""
+NEO4J_CREATE_NODE_SET_PART = "SET node.`{attr_name}` = {{`{attr_name}`}}"
+NEO4J_CREATE_NODE_SET_PART_MERGE_ATTR = "SET node.`{attr_name}` = (CASE WHEN not exists(node.`{attr_name}`) THEN [{{`{attr_name}`}}] ELSE node.`{attr_name}` + {{`{attr_name}`}}  END)"
+NEO4J_CREATE_NODE_MERGE_CHECK_PART = """WITH CASE WHEN check = 0 THEN (CASE  WHEN EXISTS(node.`{attr_name}`) AND node.`{attr_name}` <> {{`{attr_name}`}} THEN 1 ELSE 0 END) ELSE 1 END as check , node"""
+NEO4J_CREATE_NODE_RETURN_ID = """MERGE (node:{label} {{{formal_attributes}}})
+                                WITH 0 as check, node
+                                {merge_check_statement}
+                                {set_statement}
+                                RETURN ID(node) as ID, check """  # args: provType, values
+NEO4J_CREATE_RELATION_RETURN_ID = """
+                                MATCH
+                                    (from{{`meta:identifier`:'{from_identifier}'}}),
+                                    (to{{`meta:identifier`:'{to_identifier}'}})
+                                MERGE
+                                    (from)-[r:{relation_type} {{{formal_attributes}}}]->(to)
+                                    WITH 0 as check, r as node
+                                    {merge_check_statement}
+                                    {set_statement}
+                                RETURN
+                                    ID(node) as ID, check
+                                """  # args: provType, values
+# get
+NEO4j_GET_BUNDLE_RETURN_BUNDLE_NODE = """
+                        MATCH (b {`meta:prov_type`:'prov:Bundle'}) WHERE ID(b)=toInt({bundle_id}) RETURN b
+                    """
+NEO4J_Get_BUNDLES_RETURN_BUNDLE_IDS = """
+                        MATCH (d {`meta:parent_id`:{parent_id}, `meta:prov_type`: 'prov:Bundle'}) Return id(d) as ID
+                    """
+NEO4J_GET_RECORDS_BY_PROPERTY_DICT = """
+                            MATCH (d {{{filter_dict}}} )-[r]-(x {{{filter_dict}}})
+                            RETURN DISTINCT r as re
+                            //Get all nodes that are alone without connections to other nodes
+                            UNION
+                            MATCH (a {{{filter_dict}}})
+                            RETURN DISTINCT a as re
+                        """
+NEO4J_GET_RECORDS_TAIL_BY_FILTER = """
+                            MATCH (x {{{filter_dict}}})-[r *{depth}]-(y)
+                            RETURN  DISTINCT y as re
+                            UNION
+                            MATCH (x {{{filter_dict}}})-[r *{depth}]-(y)
+                            WITH REDUCE(output = [], r IN r | output + r) AS flat
+                            UNWIND flat as re
+                            RETURN DISTINCT re
+                        """
+
+NEO4J_GET_BUNDLE_RECORDS = """ 
+                            MATCH (x {`meta:identifier`: {`meta:identifier`}})-[r *1]-(y)
+                            WHERE ALL (rel in r WHERE rel.`prov:type` = 'prov:bundleAssociation')
+                            RETURN  DISTINCT y as re
+                            UNION
+                            //get all relations between the nodes
+                            MATCH (origin {`meta:identifier`: {`meta:identifier`}})-[r *1]-(x)-[r_return *1]-(y)-[r_2 *1]-(origin {`meta:identifier`: {`meta:identifier`}})
+                            WHERE ALL (rel in r WHERE rel.`prov:type` = 'prov:bundleAssociation')
+                            AND ALL (rel in r_2 WHERE rel.`prov:type` = 'prov:bundleAssociation')
+                            WITH REDUCE(output = [], r IN r_return | output + r) AS flat
+                            UNWIND flat as re
+                            RETURN DISTINCT re
+                            //get all mentionof relations
+                            UNION
+                            MATCH (bundle_1 {`meta:identifier`: {`meta:identifier`}})-[r *1]-(x)-[r_return *1]-(y)-[r_2 *1]-(bundle_2)
+                            WHERE ALL (rel in r WHERE rel.`prov:type` = 'prov:bundleAssociation')
+                            AND ALL (rel in r_2 WHERE rel.`prov:type` = 'prov:bundleAssociation')
+                            AND ALL (rel in r_return WHERE rel.`meta:prov_type` = 'prov:Mention'  and startNode(rel) = x)
+                            WITH REDUCE(output = [], r IN r_return | output + r) AS flat
+                            UNWIND flat as re
+                            RETURN DISTINCT re
+
+ """
+
+NEO4J_GET_RECORD_RETURN_NODE = """MATCH (node) WHERE ID(node)={record_id} RETURN node"""
+NEO4J_GET_RELATION_RETURN_NODE = """MATCH ()-[relation]-() WHERE ID(relation)={relation_id}  RETURN relation"""
+
+# delete
+NEO4J_DELETE__NODE_BY_ID = """MATCH (x) Where ID(x) = {node_id} DETACH DELETE x """
+NEO4J_DELETE_NODE_BY_PROPERTIES = """MATCH (n {{{filter_dict}}}) DETACH DELETE n"""
+NEO4J_DELETE_BUNDLE_NODE_BY_ID = """MATCH (b) WHERE id(b)=toInt({bundle_id}) DELETE b """
+NEO4J_DELETE_RELATION_BY_ID = """MATCH ()-[r]-() WHERE id(r) = {relation_id} DELETE r"""
